requirements.txt
# Core dependencies  
torch>=2.0.0  
torchvision>=0.15.0  
numpy>=1.21.0  

# Computer vision  
opencv-python>=4.8.0  
pillow>=9.0.0  
matplotlib>=3.5.0  
scikit-image>=0.19.0  

# Scientific computing  
scipy>=1.9.0  

# Configuration  
yacs>=0.1.8  

# Logging and visualization  
tensorboard>=2.13.0  
wandb>=0.15.0  
tqdm>=4.65.0  

# Data processing  
pandas>=1.5.0  
h5py>=3.8.0  
pycocotools>=2.0.6  

# Video processing  
imageio>=2.31.0  
imageio-ffmpeg>=0.4.8  
decord>=0.6.0  

# Utilities  
pyyaml>=6.0  
termcolor>=2.3.0  
tabulate>=0.9.0  
fvcore>=0.1.5  

# Optional: For SAM integration  
# segment-anything @ git+https://github.com/facebookresearch/segment-anything.git  

# Optional: For DINOv2  
# git+https://github.com/facebookresearch/dinov2.git  

# Development  
pytest>=7.4.0  
black>=23.7.0  
flake8>=6.1.0  
isort>=5.12.0  